Directions v1.0
Fixed Collisions
(old code commented out)

Also shifted colours to new Materials folder inside the Graphics folder
