+++ Assets/Experiment.unity
@@ -331,7 +331,7 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: 09d83ed304791664ba2c113f6d66bcee, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  player: {fileID: 0}
+  player: {fileID: 319132274}
 --- !u!1 &602047103
 GameObject:
   m_ObjectHideFlags: 0
@@ -343,9 +343,7 @@ GameObject:
   - component: {fileID: 602047108}
   - component: {fileID: 602047107}
   - component: {fileID: 602047106}
-  - component: {fileID: 602047105}
   - component: {fileID: 602047109}
-  - component: {fileID: 602047111}
   - component: {fileID: 602047110}
   m_Layer: 0
   m_Name: Player
@@ -354,19 +352,6 @@ GameObject:
   m_NavMeshLayer: 0
   m_StaticEditorFlags: 0
   m_IsActive: 1
---- !u!65 &602047105
-BoxCollider:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 602047103}
-  m_Material: {fileID: 0}
-  m_IsTrigger: 0
-  m_Enabled: 1
-  serializedVersion: 2
-  m_Size: {x: 1, y: 1, z: 1}
-  m_Center: {x: 0, y: 0, z: 0}
 --- !u!23 &602047106
 MeshRenderer:
   m_ObjectHideFlags: 0
@@ -444,9 +429,8 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: 7f69fd2f390b21d4e8a16a1c8034ec58, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  playerobject: {fileID: 602047103}
-  speed: 5
-  turnningspeed: 20
+  controller: {fileID: 602047110}
+  SPEED: 5
 --- !u!143 &602047110
 CharacterController:
   m_ObjectHideFlags: 0
@@ -465,20 +449,6 @@ CharacterController:
   m_SkinWidth: 0.08
   m_MinMoveDistance: 0.001
   m_Center: {x: 0, y: 0, z: 0}
---- !u!114 &602047111
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 602047103}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 25891c93d46c0ca44b21a397e96eb95c, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
-  controller: {fileID: 602047110}
-  SPEED: 5
 --- !u!1 &1018121878
 GameObject:
   m_ObjectHideFlags: 0
