--- Assets/Scripts (Controller)/SnakeBehaviour/PlayerMovement.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 25891c93d46c0ca44b21a397e96eb95c
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
